requirements.txt
matplotlib
ipython
graphviz
scikit-learn
pandas
geopandas
scipy
tqdm
pydot
laspy
lazrs
tensorflow-gpu==2.10